wizard update

+++ install/EE_wizard.sh
@@ -5,17 +5,19 @@ set -euo pipefail
 
 
 # Bash menu script for hypernets_tools installation.
-echo 
-echo "  _   _                                  _"
-echo " | | | |_   _ _ __   ___ _ __ _ __   ___| |_ ___"
-echo " | |_| | | | | '_ \ / _ \ '__| '_ \ / _ \ __/ __|"
-echo " |  _  | |_| | |_) |  __/ |  | | | |  __/ |_\__ \ "
-echo " |_| |_|\__, | .__/ \___|_|  |_| |_|\___|\__|___/"
-echo "        |___/|_|"
-echo 
-echo "This script aims to help hypernets_tools installation"
-echo "(Yoctopuce USB mode only)"
-echo 
+function print_logo(){
+	echo 
+	echo "  _   _                                  _"
+	echo " | | | |_   _ _ __   ___ _ __ _ __   ___| |_ ___"
+	echo " | |_| | | | | '_ \ / _ \ '__| '_ \ / _ \ __/ __|"
+	echo " |  _  | |_| | |_) |  __/ |  | | | |  __/ |_\__ \ "
+	echo " |_| |_|\__, | .__/ \___|_|  |_| |_|\___|\__|___/"
+	echo "        |___/|_|"
+	echo 
+	echo "This script aims to help hypernets_tools installation"
+	echo "(Yoctopuce USB mode only)"
+	echo 
+}
 
 
 function check_sudo_user(){
@@ -37,17 +39,11 @@ function check_if_online(){
 }
 
 
-function download_repo(){
-	# if [ -d "hypernets_tools" ]; then
-	# 	echo "The hypernets_tools already folder exists."
-	# 	exit 1
-	# fi
-
+function update_repo(){
 	echo 
 	echo 
-	# echo "Step 1 -- Download and Update Hypernets Tools..."
+	echo "-- Update Hypernets Tools..."
 	echo "------------------------------------------------"
-	# sudo -u $SUDO_USER git clone https://github.com/hypernets/hypernets_tools 
 	sudo -u $SUDO_USER git checkout beta
 	sudo -u $SUDO_USER git pull
 }
@@ -56,36 +52,43 @@ function download_repo(){
 function download_yoctohub(){
 	echo 
 	echo 
-	echo "Step 2 -- Download and Install the Yoctohub..."
+	echo "-- Download and Install the Yoctohub..."
 	echo "------------------------------------------------"
 	sudo ./install/00_install_yoctohub.sh
 }
 
 function auto_config_yocto(){
+	echo 
+	echo 
+	echo "-- Auto-config for YoctoPictor..."
+	echo "------------------------------------------------"
+
+	if [[ -f "config_static.ini" ]] || [[ -f "config_dynamic.ini" ]]; then
+		echo "Error: configuration files found, please remove it first."
+		return
+	fi
+
+	echo "Please connect the Yocto-Pictor from the 'config port' and pressenter to continue"
+	read
+
 	echo "Copying configuration files"
 	sudo -u $SUDO_USER cp hypernets/resources/config_static.ini.template config_static.ini
 	sudo -u $SUDO_USER cp hypernets/resources/config_dynamic.ini.template config_dynamic.ini
 
-	echo 
 	echo 
 	echo "Running auto config for config_static.ini..."
 
 	json_api=$(wget -O- http://127.0.0.1:4444/api.json 2> /dev/null)
 
 	yocto_id2=$(echo $json_api | python3 -c \
-		"import sys, json;
-	print(json.load(sys.stdin)['services']['whitePages'][1]['serialNumber'])")
+		"import sys, json; print(json.load(sys.stdin)['services']['whitePages'][1]['serialNumber'])")
 
 	yocto_gps=$(echo $json_api | python3 -c \
-		"import sys, json;
-	print(json.load(sys.stdin)['services']['yellowPages']['HubPort'][0]['logicalName'])")
+		"import sys, json; print(json.load(sys.stdin)['services']['yellowPages']['HubPort'][0]['logicalName'])")
 
 	yocto_id1=$(echo $json_api | python3 -c \
-		"import sys, json;
-	print(json.load(sys.stdin)['services']['yellowPages']['HubPort'][1]['logicalName'])")
+		"import sys, json; print(json.load(sys.stdin)['services']['yellowPages']['HubPort'][1]['logicalName'])")
 
-	echo 
-	echo 
 	echo "Yocto IDs are : $yocto_id1, $yocto_id2 and $yocto_gps" 
 
 	sudo -u $SUDO_USER sed -i -e '/OBSVLFR1/s/XXXXXX/'${yocto_id1:9:6}'/' config_static.ini
@@ -96,28 +99,44 @@ function auto_config_yocto(){
 function install_dependencies(){
 	echo 
 	echo 
-	echo "Step 3 -- Installation of python dependencies..."
+	echo "-- Installation of python dependencies..."
 	echo "------------------------------------------------"
-	sudo -u $SUDO_USER ./install/01_dependencies.sh
+	sudo ./install/01_dependencies.sh
 }
+
+
+function update_libhypstar(){
+	echo 
+	echo 
+	echo "-- Downloading and installing libhypstar..."
+	echo "------------------------------------------------"
+	sudo ./install/03_update_libhypstar.sh 
+}
+
 
 function configure_port(){
 	echo 
 	echo 
-	echo "Step 4 -- Configuration of the Hypstar Port..."
+	echo "-- Configuration of the Hypstar Port..."
 	echo "------------------------------------------------"
+	echo "Please connect the FTDI card and press enter to continue"
+	read
 	sudo ./install/02_configure_ports.sh
 }
 
+
 function main_menu(){
 
+	echo "------------------------------------------------"
 	PS3='Please select an option:'
 	options=(
 		"Update hypernets_tools"
+ 		"Install Dependencies"
 		"Download and install YoctoHub" 
  		"Run Yocto-Pictor auto-configuration"
- 		"Install Dependencies"
+		"Install / Update libhypstar"
 		"Configure Hypstar Port"
+		# "Check installation before field deployment"
  		"Quit")
 
 	select opt in "${options[@]}"
@@ -125,38 +144,44 @@ function main_menu(){
 		case $opt in 
 			"${options[0]}")
 				check_if_online
-				download_repo
+				update_repo
+				break
 				;;
 			"${options[1]}")
 				check_if_online
-				download_yoctohub
+				install_dependencies
+				break
 				;;
 			"${options[2]}")
-				auto_config_yocto
+				check_if_online
+				download_yoctohub
+				break
 				;;
 			"${options[3]}")
-				check_if_online
-				install_dependencies
+				auto_config_yocto
+				break
 				;;
 			"${options[4]}")
-				echo "Not implemented!"
+				check_if_online
+				update_libhypstar
+				break
 				;;
-				*) echo "Invalid choice!"
+			"${options[5]}")
+				configure_port
+				break
+				;;
+			"${options[6]}")
+				echo "Not implemented!"
+				break
+				;;
+			*) 
+				echo "Invalid choice!"
+				;;
 		esac
 	done
 }
 
+# TODO: error handler
+print_logo
 check_sudo_user
 main_menu
-# install_dependencies
-
-
-echo 
-echo 
-echo "Step 5 -- Downloading and installing libhypstar..."
-echo "------------------------------------------------"
-# sudo ./install/03_update_libhypstar.sh 
-
-echo 
-echo 
-echo "Installation Complete !"
